client/requirements.txt
speedtest-cli==2.1.3
httpx==0.21.2
grpcio-tools==1.43.0